Extract GeocoderResponse from Geocoder class
