+++ app/Services/Geocoder/GeocoderResponse.php
@@ -1,29 +1,13 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\Geocoder;
 
 use App\Casts\Coordinates;
-use App\Exceptions\InvalidAddressGeocodingException;
-use Illuminate\Support\Facades\Http;
 
-class Geocoder
+class GeocoderResponse
 {
-    private $response;
-
-    public function geocode(string $address): self
+    public function __construct(protected $response)
     {
-        if ($this->isInvalidAddress($address)) {
-            throw new InvalidAddressGeocodingException;
-        }
-
-        $this->response = $this->requestGeocoding($address);
-
-        if (! count($this->response['results'])) {
-            cache()->set('invalid-address::' . md5($address), true);
-            throw new InvalidAddressGeocodingException;
-        }
-
-        return $this;
     }
 
     public function getCoordinates(): Coordinates
@@ -44,19 +28,6 @@ class Geocoder
             : [$city, $country];
 
         return collect($values)->filter()->implode(', ');
-    }
-
-    private function requestGeocoding($address)
-    {
-        return Http::acceptJson()
-            ->withHeaders([
-                'User-Agent' => 'Symposium CLI',
-            ])
-            ->get('https://maps.googleapis.com/maps/api/geocode/json', [
-                'address' => $address,
-                'key' => config('services.google.maps.key'),
-            ])
-            ->json();
     }
 
     private function getCoordinate($type)
@@ -83,10 +54,5 @@ class Geocoder
     {
         return collect(data_get($this->response, 'results.0.address_components', []))
             ->firstWhere(fn ($component) => collect($component['types'])->intersect($types)->isNotEmpty());
-    }
-
-    private function isInvalidAddress($address)
-    {
-        return cache('invalid-address::' . md5($address));
     }
 }
